misc: rename components and give contextual name to netsuite ones
